refactor: use web-vitals for performance metrics
