+++ src/runtime/plugins/web-vitals/utils.ts
